dex_scripts/requirements.txt
pillow==12.1.0
pyelftools==0.32
PyYAML==6.0.3
